Controller Methods

+++ MEDAPP.Services/IPatientService.cs
@@ -9,7 +9,7 @@ namespace MEDAPP.Services
     public interface IPatientService
     {
         Task<List<Patient>> FindAll<T>();
-        Task<Patient> FindById<T>(long id);
+        Task<Patient> FindById<T>(int id);
         Task CreateAsync<T>(Patient entity);
         Task UpdateAsync<T>(Patient entity);
         Task DeleteAsync<T>(Patient entity);
